This branch has all of the unit test debug switches turned on.  It works but....
There are still issues with the default values created in webworkClient and passed to renderProblem in xmlrpcCall(input).

There are still warning messages at 1166 of Translator.pm and elsewhere but the problems now at least render. (The problem was caused by the changing of inputs_ref to input_hash -- there was a mismatch.)